pyproject.toml
[tool.poetry]
name = "plummet"
version = "0.1.0"
description = "Roughtime interop harness"
license = "Apache-2.0"
authors = ["Sarah Grant <sarah.grant.ietf@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pyyaml = "^6.0.2"

[tool.poetry.scripts]
plummet = 'plummet.cli:main'

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"